pyproject.toml
[tool.poetry]
name = "bkw-email-automation"
version = "0.1.0"
description = "BKW email automation: SharePoint data extraction, transformation, and email delivery"
authors = ["Your Name <your.email@company.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "bkw_automation"}]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.0.0"
openpyxl = "^3.10.0"
office365-rest-python-client = "^2.5.0"
azure-identity = "^1.14.0"
azure-storage-blob = "^12.18.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
jupyterlab = "^4.0.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"
ipykernel = "^6.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
